Move configuration code to separate file
